backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-dotenv==1.0.1
httpx==0.28.1
authlib==1.4.1
itsdangerous==2.2.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
python-multipart==0.0.19
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
tzdata>=2024.1
